refactor(reservation-application): Room 패키지에서 Seat 기능을 패키지로 분리
