--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/CreateSeatUseCase.java
@@ -1,8 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in;
-
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.CreateSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.result.SeatResult;
-
-public interface CreateSeatUseCase {
-    SeatResult create(CreateSeatCommand command);
-}
