--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/DeleteSeatUseCase.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in;
-
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.DeleteSeatCommand;
-
-public interface DeleteSeatUseCase {
-    void delete(DeleteSeatCommand command);
-}
