--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/FindSeatUseCase.java
@@ -1,8 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in;
-
-import com.seatliberator.seatliberator.reservation.application.room.port.in.query.FindSeatQuery;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.result.SeatResult;
-
-public interface FindSeatUseCase {
-    SeatResult find(FindSeatQuery query);
-}
