--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/MoveSeatUseCase.java
@@ -1,8 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in;
-
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.MoveSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.result.SeatResult;
-
-public interface MoveSeatUseCase {
-    SeatResult move(MoveSeatCommand command);
-}
