--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/query/FindSeatQuery.java
@@ -1,7 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in.query;
-
-public record FindSeatQuery(
-        String roomId,
-        String seatId
-) {
-}
