--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/port/in/query/ListSeatQuery.java
@@ -1,6 +1,0 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in.query;
-
-public record ListSeatQuery(
-        String roomId
-) {
-}
