+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/CreateSeatUseCase.java
@@ -1,0 +1,8 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in;
+
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.CreateSeatCommand;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.result.SeatResult;
+
+public interface CreateSeatUseCase {
+    SeatResult create(CreateSeatCommand command);
+}
