+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/DeleteSeatUseCase.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in;
+
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.DeleteSeatCommand;
+
+public interface DeleteSeatUseCase {
+    void delete(DeleteSeatCommand command);
+}
