+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/FindSeatUseCase.java
@@ -1,0 +1,8 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in;
+
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.query.FindSeatQuery;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.result.SeatResult;
+
+public interface FindSeatUseCase {
+    SeatResult find(FindSeatQuery query);
+}
