+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/ListSeatUseCase.java
@@ -1,0 +1,10 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in;
+
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.query.ListSeatQuery;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.result.SeatResult;
+
+import java.util.List;
+
+public interface ListSeatUseCase {
+    List<SeatResult> list(ListSeatQuery query);
+}
