+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/command/CreateSeatCommand.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in.command;
+
+public record CreateSeatCommand(
+        String roomId,
+        String seatId
+) {
+}
