+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/query/FindSeatQuery.java
@@ -1,0 +1,7 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in.query;
+
+public record FindSeatQuery(
+        String roomId,
+        String seatId
+) {
+}
