+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/query/ListSeatQuery.java
@@ -1,0 +1,6 @@
+package com.seatliberator.seatliberator.reservation.application.seat.port.in.query;
+
+public record ListSeatQuery(
+        String roomId
+) {
+}
