+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/port/in/result/SeatResult.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.application.room.port.in.result;
+package com.seatliberator.seatliberator.reservation.application.seat.port.in.result;
 
 import com.seatliberator.seatliberator.reservation.domain.seat.Seat;
 import com.seatliberator.seatliberator.reservation.domain.seat.SeatStatus;
