+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/seat/service/SeatCommandService.java
@@ -1,17 +1,17 @@
-package com.seatliberator.seatliberator.reservation.application.room.service;
+package com.seatliberator.seatliberator.reservation.application.seat.service;
 
-import com.seatliberator.seatliberator.reservation.application.room.internal.SeatAssignmentService;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.CreateSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.DeleteSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.MoveSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.UpdateSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.CreateSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.DeleteSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.MoveSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.command.UpdateSeatCommand;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.result.SeatResult;
-import com.seatliberator.seatliberator.reservation.application.room.port.out.SeatReader;
-import com.seatliberator.seatliberator.reservation.application.room.port.out.SeatStore;
+import com.seatliberator.seatliberator.reservation.application.seat.internal.SeatAssignmentService;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.CreateSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.DeleteSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.MoveSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.UpdateSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.CreateSeatCommand;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.DeleteSeatCommand;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.MoveSeatCommand;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.command.UpdateSeatCommand;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.result.SeatResult;
+import com.seatliberator.seatliberator.reservation.application.seat.port.out.SeatReader;
+import com.seatliberator.seatliberator.reservation.application.seat.port.out.SeatStore;
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationErrorCode;
 import com.seatliberator.seatliberator.reservation.application.shared.exception.ReservationApplicationException;
 import com.seatliberator.seatliberator.reservation.domain.shared.SimpleSeatLocator;
