+++ reservation/reservation-web-mvc/src/test/java/com/seatliberator/seatliberator/reservation/infrastructure/web/room/SeatQueryControllerMvcTest.java
@@ -1,10 +1,10 @@
 package com.seatliberator.seatliberator.reservation.infrastructure.web.room;
 
-import com.seatliberator.seatliberator.reservation.application.room.port.in.FindSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.ListSeatUseCase;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.query.FindSeatQuery;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.query.ListSeatQuery;
-import com.seatliberator.seatliberator.reservation.application.room.port.in.result.SeatResult;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.FindSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.ListSeatUseCase;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.query.FindSeatQuery;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.query.ListSeatQuery;
+import com.seatliberator.seatliberator.reservation.application.seat.port.in.result.SeatResult;
 import com.seatliberator.seatliberator.reservation.domain.seat.SeatStatus;
 import com.seatliberator.seatliberator.reservation.infrastructure.web.room.controller.SeatQueryController;
 import org.junit.jupiter.api.BeforeEach;
